the logic of checks has been moved to the service for users and places

+++ src/main/java/com/events/eventPlanner/service/UserService.java
@@ -4,16 +4,20 @@ import com.events.eventPlanner.domain.DTO.EventDbDto;
 import com.events.eventPlanner.domain.DTO.UserResponseDto;
 import com.events.eventPlanner.domain.Event;
 import com.events.eventPlanner.domain.User;
+import com.events.eventPlanner.exceptions.ForbiddenContentException;
+import com.events.eventPlanner.exceptions.ObjectNotFoundException;
 import com.events.eventPlanner.mappers.DtoMapper;
 import com.events.eventPlanner.repository.EventRepository;
 import com.events.eventPlanner.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.sql.Date;
 import java.util.ArrayList;
+import java.util.Objects;
 
 @Service
 public class UserService {
@@ -30,21 +34,36 @@ public class UserService {
     }
 
     public UserResponseDto getUserById(int id) {
-        User user = userRepository.findById(id).orElseThrow();
+        User user = userRepository.findById(id).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find user with id: " + id));
+        checkPermission(id);
         return DtoMapper.fromUserToUserResponseDto(user);
-    }
-
-    public User findUserByLogin(String login) {
-        return userRepository.findByLogin(login).orElse(null);
     }
 
     public ArrayList<UserResponseDto> getAllUsers() {
         ArrayList<User> users = (ArrayList<User>) userRepository.findAll();
-        ArrayList<UserResponseDto> userResponseDtos = new ArrayList<>();
+        ArrayList<UserResponseDto> usersResponseDto = new ArrayList<>();
         for (User u : users) {
-            userResponseDtos.add(DtoMapper.fromUserToUserResponseDto(u));
+            usersResponseDto.add(DtoMapper.fromUserToUserResponseDto(u));
         }
-        return userResponseDtos;
+        if (usersResponseDto.isEmpty()) {
+            throw new ObjectNotFoundException("Don't find any users");
+        }
+        return usersResponseDto;
+    }
+
+    public ArrayList<Event> getAllEventsForUser(int id) {
+        userRepository.findById(id).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find user with id: " + id));
+        checkPermission(id);
+        ArrayList<Event> events = new ArrayList<>();
+        for (EventDbDto e : eventRepository.getAllEventsForUser(id)) {
+            events.add(DtoMapper.fromEventDbDtoToEvent(e));
+        }
+        if (events.isEmpty()) {
+            throw new ObjectNotFoundException("No events for this user");
+        }
+        return events;
     }
 
     public User createUser(User user) {
@@ -56,32 +75,51 @@ public class UserService {
     }
 
     @Transactional
-    public int addEventToUser(int eventID, int userID) {
+    public void addEventToUser(int eventId, int userId) {
+        userRepository.findById(userId).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find user with id: " + userId));
+        eventRepository.findById(eventId).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find event with id: " + eventId));
+        checkPermission(userId);
         Date createTime = new Date(System.currentTimeMillis());
-        return userRepository.addEventToUser(eventID, userID, createTime);
+        userRepository.addEventToUser(eventId, userId, createTime);
     }
 
-    public ArrayList<Event> getAllEventsForUser(int userId) {
-        ArrayList<Event> events = new ArrayList<>();
-        for (EventDbDto e : eventRepository.getAllEventsForUser(userId)) {
-            events.add(DtoMapper.fromEventDbDtoToEvent(e));
-        }
-        return events;
-    }
-
-    @Transactional
-    public int deleteEventFromUser(int eventID, int userID) {
-        return userRepository.deleteEventFromUser(eventID, userID);
-    }
-
-    public User updateUser(User user) {
+    public void updateUser(User user) {
+        userRepository.findById(user.getId()).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find user with id: " + user.getId()));
+        checkPermission(user.getId());
         user.setEdited(new Date(System.currentTimeMillis()));
         user.setPassword(passwordEncoder.encode(user.getPassword()));
-        return userRepository.save(user);
+        userRepository.save(user);
     }
 
     @Transactional
     public void deleteUser(int id) {
+        userRepository.findById(id).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find user with id: " + id));
+        checkPermission(id);
         userRepository.deleteUser(id);
+    }
+
+    @Transactional
+    public void deleteEventFromUser(int eventId, int userId) {
+        userRepository.findById(userId).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find user with id: " + userId));
+        eventRepository.findById(eventId).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find event with id: " + eventId));
+        checkPermission(userId);
+        userRepository.deleteEventFromUser(eventId, userId);
+    }
+
+    public void checkPermission(int id) {
+        String login = SecurityContextHolder.getContext().getAuthentication().getName();
+        User user = userRepository.findByLogin(login).orElseThrow(
+                () -> new ObjectNotFoundException("Don't find user in secure context"));
+        if (!Objects.equals(user.getRole(), "admin")) {
+            if (!user.getId().equals(id)) {
+                throw new ForbiddenContentException("You are not this user");
+            }
+        }
     }
 }
